+++ crates/matrix-sdk-ui/src/timeline/event_handler.rs
@@ -31,7 +31,8 @@ use ruma::{
         receipt::Receipt,
         relation::Replacement,
         room::message::{
-            Relation, RoomMessageEventContent, RoomMessageEventContentWithoutRelation,
+            Relation, RelationWithoutReplacement, RoomMessageEventContent,
+            RoomMessageEventContentWithoutRelation,
         },
         AnyMessageLikeEventContent, AnySyncMessageLikeEvent, AnySyncStateEvent,
         AnySyncTimelineEvent, BundledMessageLikeRelations, EventContent, FullStateEventContent,
@@ -147,6 +148,7 @@ pub(super) enum HandleAggregationKind {
 
 /// An action that we want to cause on the timeline.
 #[derive(Clone, Debug)]
+#[allow(clippy::large_enum_variant)]
 pub(super) enum TimelineAction {
     /// Add a new timeline item.
     ///
@@ -354,11 +356,44 @@ impl TimelineAction {
             }
 
             AnyMessageLikeEventContent::Sticker(content) => {
+                let mut replied_to_event_id = None;
+                let mut thread_root = None;
+                let in_reply_to_details =
+                    content.relates_to.as_ref().and_then(|relation| match relation {
+                        Relation::Reply { in_reply_to } => {
+                            replied_to_event_id = Some(in_reply_to.event_id.clone());
+                            Some(InReplyToDetails::new(
+                                in_reply_to.event_id.clone(),
+                                timeline_items,
+                            ))
+                        }
+                        Relation::Thread(thread) => {
+                            thread_root = Some(thread.event_id.clone());
+                            thread.in_reply_to.as_ref().map(|in_reply_to| {
+                                replied_to_event_id = Some(in_reply_to.event_id.clone());
+                                InReplyToDetails::new(in_reply_to.event_id.clone(), timeline_items)
+                            })
+                        }
+                        _ => None,
+                    });
+
+                // If this message is a reply to another message, add an entry in the
+                // inverted mapping.
+                if let Some(event_id) = event_id {
+                    if let Some(replied_to_event_id) = replied_to_event_id {
+                        // This is a reply! Add an entry.
+                        meta.replies
+                            .entry(replied_to_event_id)
+                            .or_default()
+                            .insert(event_id.to_owned());
+                    }
+                }
+
                 Self::add_item(TimelineItemContent::MsgLike(MsgLikeContent {
                     kind: MsgLikeKind::Sticker(Sticker { content }),
                     reactions: Default::default(),
-                    thread_root: None,
-                    in_reply_to: None,
+                    thread_root,
+                    in_reply_to: in_reply_to_details,
                     thread_summary: None,
                 }))
             }
@@ -393,6 +428,39 @@ impl TimelineAction {
                     .or(pending_edit)
                     .unzip();
 
+                let mut replied_to_event_id = None;
+                let mut thread_root = None;
+                let in_reply_to_details =
+                    c.relates_to.as_ref().and_then(|relation| match relation {
+                        RelationWithoutReplacement::Reply { in_reply_to } => {
+                            replied_to_event_id = Some(in_reply_to.event_id.clone());
+                            Some(InReplyToDetails::new(
+                                in_reply_to.event_id.clone(),
+                                timeline_items,
+                            ))
+                        }
+                        RelationWithoutReplacement::Thread(thread) => {
+                            thread_root = Some(thread.event_id.clone());
+                            thread.in_reply_to.as_ref().map(|in_reply_to| {
+                                replied_to_event_id = Some(in_reply_to.event_id.clone());
+                                InReplyToDetails::new(in_reply_to.event_id.clone(), timeline_items)
+                            })
+                        }
+                        _ => None,
+                    });
+
+                // If this message is a reply to another message, add an entry in the
+                // inverted mapping.
+                if let Some(event_id) = event_id {
+                    if let Some(replied_to_event_id) = replied_to_event_id {
+                        // This is a reply! Add an entry.
+                        meta.replies
+                            .entry(replied_to_event_id)
+                            .or_default()
+                            .insert(event_id.to_owned());
+                    }
+                }
+
                 let poll_state = PollState::new(c, edit_content);
 
                 let edit_json = edit_json.flatten();
@@ -401,8 +469,8 @@ impl TimelineAction {
                     content: TimelineItemContent::MsgLike(MsgLikeContent {
                         kind: MsgLikeKind::Poll(poll_state),
                         reactions: Default::default(),
-                        thread_root: None,
-                        in_reply_to: None,
+                        thread_root,
+                        in_reply_to: in_reply_to_details,
                         thread_summary: None,
                     }),
                     edit_json,